+++ packages/realm/src/Configuration.ts
@@ -157,6 +157,12 @@ export type Configuration = {
    */
   onFirstOpen?: (realm: Realm) => void;
   migrationOptions?: MigrationOptions;
+  /**
+   * Specifies if this Realm should be excluded from iCloud backup.
+   * @default false
+   * @since 12.13.3
+   */
+  excludeFromIcloudBackup?: boolean;
 };
 
 /**